pyproject.toml
[tool.poetry]
name = "sonic_exporter"
version = "0.0.0"
description = "\"Sonic Exporter\""
authors = ["Marcel Fest <marcel.fest@telekom.de>"]
license = "MIT License"

exclude = ["src/sonic_exporter/test"]

[tool.poetry.dependencies]
python = "^3.10"
prometheus-client = "^0.15.0"
dacite = "^1.6.0"
PyYAML = "^6.0"
jc = "^1.22.1"

[tool.poetry.dev-dependencies]
black = "*"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}
mypy = "^1.7.1"
types-pyyaml = "^6.0.12.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sonic_exporter = 'sonic_exporter.exporter:cli'
